+++ funding/get_funded_transaction.js
@@ -25,6 +25,7 @@ const {isArray} = Array;
   {
     id: <Transaction Id Hex String>
     [inputs]: [{
+      [lock_expires_at]: <UTXO Lock Expires At ISO 8601 Date String>
       [lock_id]: <UTXO Lock Id Hex String>
       transaction_id: <Transaction Hex Id String>
       transaction_vout: <Transaction Output Index Number>