+++ funding/index.js
@@ -1,5 +1,11 @@
 const askForFeeRate = require('./ask_for_fee_rate');
 const getFundedTransaction = require('./get_funded_transaction');
 const getPsbtFromTransaction = require('./get_psbt_from_transaction');
+const maintainUtxoLocks = require('./maintain_utxo_locks');
 
-module.exports = {askForFeeRate, getFundedTransaction, getPsbtFromTransaction};
+module.exports = {
+  askForFeeRate,
+  getFundedTransaction,
+  getPsbtFromTransaction,
+  maintainUtxoLocks,
+};